@@ -1,3 +1,3 @@
 module github.com/aacfactory/copier
 
-go 1.17
+go 1.21
